Refactor  Board Class
Remove moves method and update moves methods, in order to simplify to
 just one array to hold the board.

 Add a win index to check for wins and a constant hash which holds the
index of each space on the board under the key for their move. That way
we can still use verbage such as 'top-left' for making moves.

Simplify the show method and modify the valid_move? method to use the
move index hash constant to check the board.

Remove unnecessary attr_readers from deleted variables and add board.