Refactor win/tie logic
Changed the way win and tie logic works. Board now has a method to check
if it is full, and has a method to check the board against the win array
Game class checks for a winner on the board, and if there is no winner
then it checks if the board is full.